Safari-611.1.10/LayoutTests/fast/forms/ios/ipad/select-should-not-bring-up-software-keyboard-expected.txt

This test verifies that tapping a select element on iPad does not bring up the software keyboard.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isShowingKeyboardAfterFocusingSelect is false
PASS isShowingKeyboardAfterFocusingColorInput is false
PASS successfullyParsed is true

TEST COMPLETE
